requirements.txt
aiofiles==25.1.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
bcrypt==4.1.3
billiard==4.2.2
celery==5.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.115.0
greenlet==3.2.4
gTTS==2.5.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
imageio==2.37.0
kombu==5.6.0
lazy_loader==0.4
lxml==6.0.2
mysql-connector-python==9.0.0
networkx==3.4.2
numpy==1.26.4
openpyxl==3.1.5
packaging==25.0
pillow==10.4.0
prompt_toolkit==3.0.52
psutil==6.0.0
pycryptodomex==3.23.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyMuPDF==1.24.10
PyMySQL==1.1.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
python-pptx==0.6.23
pyzipper==0.3.6
redis==5.0.8
requests==2.32.3
scikit-image==0.24.0
scipy==1.15.3
SQLAlchemy==2.0.35
starlette==0.38.6
tifffile==2025.5.10
tqdm==4.66.4
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.30.6
vine==5.1.0
wcwidth==0.2.14
wheel==0.45.1
xlsxwriter==3.2.9